test(datetime): migrate values
